realization optional functional - multiple config classes
