+++ hw25-di/homework/src/main/java/ru/otus/config/multiple/AppConfig1.java
@@ -1,0 +1,20 @@
+package ru.otus.config.multiple;
+
+import ru.otus.appcontainer.api.AppComponent;
+import ru.otus.appcontainer.api.AppComponentsContainerConfig;
+import ru.otus.services.*;
+
+@AppComponentsContainerConfig(order = 1)
+public class AppConfig1 {
+
+    @AppComponent(order = 0, name = "equationPreparer")
+    public EquationPreparer equationPreparer() {
+        return new EquationPreparerImpl();
+    }
+
+    @SuppressWarnings("squid:S106")
+    @AppComponent(order = 0, name = "ioService")
+    public IOService ioService() {
+        return new IOServiceStreams(System.out, System.in);
+    }
+}
