+++ hw25-di/homework/src/main/java/ru/otus/config/multiple/AppConfig2.java
@@ -1,17 +1,11 @@
-package ru.otus.config;
+package ru.otus.config.multiple;
 
 import ru.otus.appcontainer.api.AppComponent;
 import ru.otus.appcontainer.api.AppComponentsContainerConfig;
 import ru.otus.services.*;
 
 @AppComponentsContainerConfig(order = 1)
-public class AppConfig {
-
-    @AppComponent(order = 0, name = "equationPreparer")
-    public EquationPreparer equationPreparer() {
-        return new EquationPreparerImpl();
-    }
-
+public class AppConfig2 {
     @AppComponent(order = 1, name = "playerService")
     public PlayerService playerService(IOService ioService) {
         return new PlayerServiceImpl(ioService);
@@ -21,11 +15,5 @@ public class AppConfig {
     public GameProcessor gameProcessor(
             IOService ioService, PlayerService playerService, EquationPreparer equationPreparer) {
         return new GameProcessorImpl(ioService, equationPreparer, playerService);
-    }
-
-    @SuppressWarnings("squid:S106")
-    @AppComponent(order = 0, name = "ioService")
-    public IOService ioService() {
-        return new IOServiceStreams(System.out, System.in);
     }
 }
