+++ hw25-di/homework/src/main/java/ru/otus/config/single/AppConfig.java
@@ -1,0 +1,31 @@
+package ru.otus.config.single;
+
+import ru.otus.appcontainer.api.AppComponent;
+import ru.otus.appcontainer.api.AppComponentsContainerConfig;
+import ru.otus.services.*;
+
+@AppComponentsContainerConfig(order = 1)
+public class AppConfig {
+
+    @AppComponent(order = 0, name = "equationPreparer")
+    public EquationPreparer equationPreparer() {
+        return new EquationPreparerImpl();
+    }
+
+    @AppComponent(order = 1, name = "playerService")
+    public PlayerService playerService(IOService ioService) {
+        return new PlayerServiceImpl(ioService);
+    }
+
+    @AppComponent(order = 2, name = "gameProcessor")
+    public GameProcessor gameProcessor(
+            IOService ioService, PlayerService playerService, EquationPreparer equationPreparer) {
+        return new GameProcessorImpl(ioService, equationPreparer, playerService);
+    }
+
+    @SuppressWarnings("squid:S106")
+    @AppComponent(order = 0, name = "ioService")
+    public IOService ioService() {
+        return new IOServiceStreams(System.out, System.in);
+    }
+}
